+++ operator/pkg/services/authn_api.go
@@ -7,84 +7,148 @@ package services
 import (
 	"context"
 	"crypto/tls"
+	"encoding/json"
 	"fmt"
 	"net/http"
+	"os"
 	"time"
 
 	"github.com/NVIDIA/aistore/api"
 	"github.com/NVIDIA/aistore/api/authn"
-	"github.com/NVIDIA/aistore/api/env"
 	"github.com/NVIDIA/aistore/cmn"
-	"github.com/NVIDIA/aistore/cmn/cos"
+	aisv1 "github.com/ais-operator/api/v1beta1"
+	aisclient "github.com/ais-operator/pkg/client"
+	apierrors "k8s.io/apimachinery/pkg/api/errors"
+	"k8s.io/apimachinery/pkg/types"
 	logf "sigs.k8s.io/controller-runtime/pkg/log"
 )
 
-type authNConfig struct {
-	adminUser string
-	adminPass string
-	port      string
-	host      string
-	protocol  string
-}
-
 // AuthN constants
 const (
+	OperatorNamespace    = "OPERATOR_NAMESPACE"
 	AuthNServiceHostName = "ais-authn.ais"
 	AuthNServicePort     = "52001"
-	AuthNAdminUser       = "admin"
-	AuthNAdminPass       = "admin"
 
-	AuthNServiceHostVar = "AIS_AUTHN_SERVICE_HOST"
-	AuthNServicePortVar = "AIS_AUTHN_SERVICE_PORT"
+	AuthNConfigMapVar  = "AIS_AUTHN_CM"
+	AuthNSecretRefName = "SU-NAME"
+	AuthNSecretRefPass = "SU-PASS"
 )
 
 type (
 	AuthNClientInterface interface {
-		getAdminToken(ctx context.Context) (string, error)
+		getAdminToken(ctx context.Context, ais *aisv1.AIStore) (string, error)
 	}
 
 	AuthNClient struct {
-		config *authNConfig
+		k8sClient *aisclient.K8sClient
+	}
+
+	AuthNClusterConfig struct {
+		TLS             bool   `json:"tls"`
+		Host            string `json:"host"`
+		Port            string `json:"port"`
+		SecretNamespace string `json:"secretNamespace"`
+		SecretName      string `json:"secretName"`
 	}
 )
 
-func NewAuthNClient() *AuthNClient {
+func NewAuthNClient(k8sClient *aisclient.K8sClient) *AuthNClient {
 	return &AuthNClient{
-		config: newAuthNConfig(),
+		k8sClient: k8sClient,
 	}
 }
 
-// getAdminToken retrieves an admin token from AuthN service for the given AIS cluster.
-func (c AuthNClient) getAdminToken(ctx context.Context) (string, error) {
-	authNURL := fmt.Sprintf("%s://%s:%s", c.config.protocol, c.config.host, c.config.port)
-	authNBP := authNBaseParams(authNURL, "")
-	zeroDuration := time.Duration(0)
+// getAdminToken Gets an admin token for the given cluster using the credentials secret referenced by the operator's authN configmap
+func (c *AuthNClient) getAdminToken(ctx context.Context, ais *aisv1.AIStore) (string, error) {
+	authnConf, err := c.getAuthConfig(ctx, ais)
+	if err != nil || authnConf == nil || authnConf.SecretName == "" {
+		return "", err
+	}
+	secretData, err := c.getSecretData(ctx, authnConf.SecretNamespace, authnConf.SecretName)
+	if err != nil || secretData == nil {
+		return "", err
+	}
+	return getTokenFromAuthN(ctx, authNBaseParams(authnConf), secretData)
+}
 
-	tokenMsg, err := authn.LoginUser(*authNBP, c.config.adminUser, c.config.adminPass, &zeroDuration)
+// getAuthConfig Gets the data from the configmap defined by `AIS_AUTHN_CM`
+func (c *AuthNClient) getAuthConfig(ctx context.Context, ais *aisv1.AIStore) (*AuthNClusterConfig, error) {
+	logger := logf.FromContext(ctx)
+	// Get the authN credentials secret name for this cluster, if it exists
+	cmName, found := os.LookupEnv(AuthNConfigMapVar)
+	if !found {
+		return nil, nil
+	}
+	cmNs, found := os.LookupEnv(OperatorNamespace)
+	if !found {
+		logger.Info("OPERATOR_NAMESPACE environment variable not set, failed to find a ConfigMap for AuthN")
+		return nil, nil
+	}
+	cm, err := c.k8sClient.GetConfigMap(ctx, types.NamespacedName{Name: cmName, Namespace: cmNs})
+	// If the config map doesn't exist we haven't configured the operator to use authN at all
+	if err != nil && apierrors.IsNotFound(err) {
+		return nil, nil
+	}
 	if err != nil {
-		return "", fmt.Errorf("failed to login admin user to AuthN: %w", err)
+		logger.Error(err, fmt.Sprintf("Failed to get AuthN ConfigMap %s in namespace %s", cmName, cmNs))
+		return nil, err
+	}
+	if cm == nil || cm.Data == nil {
+		return nil, fmt.Errorf("AuthN ConfigMap %s in namespace %s has no data", cmName, cmNs)
+	}
+	key := ais.Namespace + "-" + ais.Name
+	confJSON, ok := cm.Data[key]
+	if !ok {
+		return nil, nil
+	}
+	var conf *AuthNClusterConfig
+	if err = json.Unmarshal([]byte(confJSON), &conf); err != nil {
+		logger.Error(err, fmt.Sprintf("Failed to unmarshal entry for cluster %s in AuthN ConfigMap %s in namespace %s", key, cmName, cmNs))
+		return nil, err
+	}
+	return conf, nil
+}
+
+// getSecretData Get the secret data from the specified secret name and namespace
+func (c *AuthNClient) getSecretData(ctx context.Context, namespace, secretName string) (map[string][]byte, error) {
+	logger := logf.FromContext(ctx)
+	// Look up the secret credentials and use them to obtain a token
+	secret, err := c.k8sClient.GetSecret(ctx, types.NamespacedName{Name: secretName, Namespace: namespace})
+	if err != nil {
+		logger.Error(err, fmt.Sprintf("Failed to get AuthN credentials secret %s in namespace %s", secretName, namespace))
+		return nil, err
+	}
+	if secret == nil || len(secret.Data) == 0 {
+		return nil, fmt.Errorf("AuthN Secret %s in namespace %s has no data", secretName, namespace)
+	}
+	return secret.Data, nil
+}
+
+// getTokenFromAuthN retrieves an admin token from AuthN using the username and password from the provided secret data
+func getTokenFromAuthN(ctx context.Context, params *api.BaseParams, secretData map[string][]byte) (string, error) {
+	logger := logf.FromContext(ctx)
+	zeroDuration := time.Duration(0)
+	user := string(secretData[AuthNSecretRefName])
+	pass := string(secretData[AuthNSecretRefPass])
+	tokenMsg, err := authn.LoginUser(*params, user, pass, &zeroDuration)
+	if err != nil {
+		return "", fmt.Errorf("failed to login %q user to AuthN: %w", user, err)
 	}
 
-	logf.FromContext(ctx).Info("Successfully logged in as Admin to AuthN")
+	logger.Info(fmt.Sprintf("Successfully fetched token for user %q from AuthN", user))
 	return tokenMsg.Token, nil
 }
 
-func newAuthNConfig() *authNConfig {
-	protocol := "http"
-	if useHTTPS, err := cos.IsParseEnvBoolOrDefault(env.AisAuthUseHTTPS, false); err == nil && useHTTPS {
-		protocol = "https"
+func authNBaseParams(conf *AuthNClusterConfig) *api.BaseParams {
+	host := conf.Host
+	port := conf.Port
+	if host == "" {
+		host = AuthNServiceHostName
+	}
+	if port == "" {
+		port = AuthNServicePort
 	}
 
-	return &authNConfig{
-		adminUser: cos.GetEnvOrDefault(env.AisAuthAdminUsername, AuthNAdminUser),
-		adminPass: cos.GetEnvOrDefault(env.AisAuthAdminPassword, AuthNAdminPass),
-		host:      cos.GetEnvOrDefault(AuthNServiceHostVar, AuthNServiceHostName),
-		port:      cos.GetEnvOrDefault(AuthNServicePortVar, AuthNServicePort),
-		protocol:  protocol,
-	}
-}
-
-func authNBaseParams(url, token string) *api.BaseParams {
 	transportArgs := cmn.TransportArgs{
 		Timeout:         10 * time.Second,
 		UseHTTPProxyEnv: true,
@@ -98,8 +162,7 @@ func authNBaseParams(url, token string) *api.BaseParams {
 			Transport: transport,
 			Timeout:   transportArgs.Timeout,
 		},
-		URL:   url,
-		Token: token,
-		UA:    userAgent,
+		URL: createAPIURL(conf.TLS, host, port),
+		UA:  userAgent,
 	}
 }